Extract RelationSmiley into its own file
